clarified comments + cleaned portions of code
