+++ variable.go
@@ -1,7 +1,5 @@
 package gal
 
-import "fmt"
-
 type Variable struct {
 	Name string
 }
@@ -18,24 +16,4 @@ func (Variable) kind() entryKind {
 
 func (v Variable) String() string {
 	return v.Name
-}
-
-type ObjectProperty struct {
-	ObjectName   string
-	PropertyName string
-}
-
-func NewObjectProperty(objectName, propertyName string) ObjectProperty {
-	return ObjectProperty{
-		ObjectName:   objectName,
-		PropertyName: propertyName,
-	}
-}
-
-func (o ObjectProperty) kind() entryKind {
-	return objectPropertyEntryKind
-}
-
-func (o ObjectProperty) String() string {
-	return fmt.Sprintf("%s.%s", o.ObjectName, o.PropertyName)
 }
